prov/efa: move and rename efa_mr_cache_init()
This patch moved efa_mr_cache_init() from efa_domain.c
to efa_mr.c, and renamed it efa_mr_cache_open() to
conform with naming convention.

The 1st argument of the function have been converted from
pointer to domain to pointer to cache.
