+++ prov/efa/src/efa_domain.c
@@ -121,122 +121,6 @@ static struct fi_ops_domain efa_domain_ops = {
 	.query_collective = fi_no_query_collective,
 };
 
-/* @brief Setup the MR cache.
- *
- * This function enables the MR cache using the util MR cache code. Note that
- * if the call to ofi_mr_cache_init fails, we continue but disable the cache.
- *
- * @param efa_domain The EFA domain where cache ops should be set
- * @param info Validated info struct selected by the user
- * @return 0 on success, fi_errno on failure.
- */
-static int efa_mr_cache_init(struct efa_domain *domain, struct fi_info *info)
-{
-	struct ofi_mem_monitor *memory_monitors[OFI_HMEM_MAX] = {
-		[FI_HMEM_SYSTEM] = default_monitor,
-		[FI_HMEM_CUDA] = cuda_monitor,
-	};
-	int ret;
-
-	/* Both Open MPI (and possibly other MPI implementations) and
-	 * Libfabric use the same live binary patching to enable memory
-	 * monitoring, but the patching technique only allows a single
-	 * "winning" patch.  The Libfabric memhooks monitor will not
-	 * overwrite a previous patch, but instead return
-	 * -FI_EALREADY.  There are three cases of concern, and in all
-	 * but one of them, we can avoid changing the default monitor.
-	 *
-	 * (1) Upper layer does not patch, such as Open MPI 4.0 and
-	 * earlier.  In this case, the default monitor will be used,
-	 * as the default monitor is either not the memhooks monitor
-	 * (because the user specified a different monitor) or the
-	 * default monitor is the memhooks monitor, but we were able
-	 * to install the patches.  We will use the default monitor in
-	 * this case.
-	 *
-	 * (2) Upper layer does patch, but does not export a memory
-	 * monitor, such as Open MPI 4.1.0 and 4.1.1.  In this case,
-	 * if the default memory monitor is not memhooks, we will use
-	 * the default monitor.  If the default monitor is memhooks,
-	 * the patch will fail to apply, and we will change the
-	 * requested monitor to UFFD to avoid a broken configuration.
-	 * If the user explicitly requested memhooks, we will return
-	 * an error, as we can not satisfy that request.
-	 *
-	 * (3) Upper layer does patch and exports a memory monitor,
-	 * such as Open MPI 4.1.2 and later.  In this case, the
-	 * default monitor will have been changed from the memhooks
-	 * monitor to the imported monitor, so we will use the
-	 * imported monitor.
-	 *
-	 * The only known cases in which we will not use the default
-	 * monitor are Open MPI 4.1.0/4.1.1.
-	 *
-	 * It is possible that this could be better handled at the
-	 * mem_monitor level in Libfabric, but so far we have not
-	 * reached agreement on how that would work.
-	 */
-	if (default_monitor == memhooks_monitor) {
-		ret = memhooks_monitor->start(memhooks_monitor);
-		if (ret == -FI_EALREADY) {
-			if (cache_params.monitor) {
-				EFA_WARN(FI_LOG_DOMAIN,
-					 "Memhooks monitor requested via FI_MR_CACHE_MONITOR, but memhooks failed to\n"
-					 "install.  No working monitor availale.\n");
-				return -FI_ENOSYS;
-			}
-			EFA_INFO(FI_LOG_DOMAIN,
-				 "Detected potential memhooks monitor conflict. Switching to UFFD.\n");
-			memory_monitors[FI_HMEM_SYSTEM] = uffd_monitor;
-		}
-	} else if (default_monitor == NULL) {
-		/* TODO: Fail if we don't find a system monitor.  This
-		 * is a debatable decision, as the VERBS provider
-		 * falls back to a no-cache mode in this case.  We
-		 * fail the domain creation because the rest of the MR
-		 * code hasn't been audited to deal with a NULL
-		 * monitor.
-		 */
-		EFA_WARN(FI_LOG_DOMAIN,
-			 "No default SYSTEM monitor available.\n");
-		return -FI_ENOSYS;
-	}
-
-	domain->cache = (struct ofi_mr_cache *)calloc(1, sizeof(struct ofi_mr_cache));
-	if (!domain->cache)
-		return -FI_ENOMEM;
-
-	if (!efa_mr_max_cached_count)
-		efa_mr_max_cached_count = info->domain_attr->mr_cnt *
-					  EFA_MR_CACHE_LIMIT_MULT;
-	if (!efa_mr_max_cached_size)
-		efa_mr_max_cached_size = domain->device->ibv_attr.max_mr_size *
-					 EFA_MR_CACHE_LIMIT_MULT;
-	/*
-	 * XXX: we're modifying a global in the util mr cache? do we need an
-	 * API here instead?
-	 */
-	cache_params.max_cnt = efa_mr_max_cached_count;
-	cache_params.max_size = efa_mr_max_cached_size;
-	domain->cache->entry_data_size = sizeof(struct efa_mr);
-	domain->cache->add_region = efa_mr_cache_entry_reg;
-	domain->cache->delete_region = efa_mr_cache_entry_dereg;
-	ret = ofi_mr_cache_init(&domain->util_domain, memory_monitors,
-				domain->cache);
-	if (!ret) {
-		domain->util_domain.domain_fid.mr = &efa_domain_mr_cache_ops;
-		EFA_INFO(FI_LOG_DOMAIN, "EFA MR cache enabled, max_cnt: %zu max_size: %zu\n",
-			 cache_params.max_cnt, cache_params.max_size);
-	} else {
-		EFA_WARN(FI_LOG_DOMAIN, "EFA MR cache init failed: %s\n",
-		         fi_strerror(ret));
-		free(domain->cache);
-		domain->cache = NULL;
-	}
-
-	return 0;
-}
-
 /* @brief Allocate a domain, open the device, and set it up based on the hints.
  *
  * This function creates a domain and uses the info struct to configure the
@@ -356,9 +240,11 @@ int efa_domain_open(struct fid_fabric *fabric_fid, struct fi_info *info,
 	 * If FI_MR_LOCAL is set, we do not want to use the MR cache.
 	 */
 	if (!app_mr_local && efa_mr_cache_enable) {
-		ret = efa_mr_cache_init(domain, info);
+		ret = efa_mr_cache_open(&domain->cache, domain);
 		if (ret)
 			goto err_free_info;
+
+		domain->util_domain.domain_fid.mr = &efa_domain_mr_cache_ops;
 	}
 
 	return 0;
